requirements.txt
torch
torchvision
Pillow
segmentation-models-pytorch
numpy
gdown
matplotlib
streamlit